Fix the console freezing upon shutting down the connected console
THAT is probably due to the shutdown MD pulling the cable low. The
cable works like this:

Set the port to output and nibble (and the one control line), set the
nibble with the control line high, wait a usec or two, then set the
control line low. If there's another MD linked and waiting, setting
the control line low made it generate an ext int and it's in the
interrupt code. It will eventually lower its own control line to show
it's got that first nibble. We see that immediately because the
machine is off and pulling the line low, but we don't know that. That
first wait has a timeout AND bumps fm. Then we raise our control line
to show we've seen that they got the nibble and wait for them to raise
their control line. However, they aren't really there and will never
raise the line. But at this point in the code, we THINK they're there
as they "responded" by lowering their control line. So this second
loop, we just busy loop waiting on the line to go high.

Had a couple other things to do that were related - the interrupt
needed to handle no console on the other side since no console =
pulled low = constant interrupt. I added a timeout in the interrupt
that turns off the networking if it fails. It doesn't seem to affect
normal networking, but fixes the no other console issue. I did
things like turn off the other console in the middle of a network game
to check. I think this is pretty tight.